Fixed an issue with Throws
For some reason, the lines that told the game to rotate the player were erroneously placed inside an if statement checking if the player is supposed to be using their knockdown sprite. For back throw, this meant that the opponent would not rotate correctly in the middle of the animation.