Fixed momentum and throw bugs
When hitting the opponent, the attack player's horizontal speed was being preserved. When hitting an opponent with standing heavy, Russel's hsp is set to 24 when the collision happens. If the player were to cancel this move with side special, that momentum would get conserved when transitioning to side special, causing him to fly across the screen.

As for throws, heldOpponent wasn't being reset when the player threw someone, causing them to override the victim's momentum data.

These were both caused by moving code around. I'm not sure what could've moved these lines, whether it was when someone was refactoring it or when the merge happened.