Added new script to process command grabs
Some command grabs may require unique code in order to work, so the ProcessCommandGrab script is responsible for the custom functionality.
For example, Russel's Up Special command grab only hits the opponent Russel touches the ground.
I'm using the move's animation timer as a way to denote the different parts of the move instead of having to make two separate moves for each part (falling and hitting the ground.)