Made Temporary fix to crash
Since Russel and Beverly don't have the same number of alternate palettes, if a player would win a game with a palette ID that was outside of another character's total number of palettes, the game would crash since the palette ID is carried over between games, and the game would be indexing a palette that doesnt exist.
To fix this, I've given Bev a bunch of Dummy Palettes.