Adjusted Cam behavior during command grabs
THe camera would shake around because the midpoint of the players was moving around. This makes sure that the game focuses on 1 player to make the camera more smooth.

Also I forgot to mention I adjusted how shadows are displayed during grabs. There was a weird glitch where the shadow would appear mid-air, so I just disabled the victim's shadow during grabs.