Polished up how blocking works
It's now always reliant on the position between players. The defending player must hold away from the attacking player. It used to work by using canBlock as a flag for when a character would block attacks, and was only set to 1 during the player's walk back and holding crouch-back. It's reworked to allow it to function in singleplayer, where the player can't walk backwards.