Implemented Gunther's standing normals
The following moves have been implemented using the character editior (this shouldn't have taken 3 hours AAHHHHHHHHHHHH)

-idle
-stand light 1
--4 frame jab
-stand light 2
--autocombo follow up that has some launch power if you somehow connect it with an airborne opponent
-stand medium
--long range poke. has two hitboxes (they don't combo into each other)
-stand heavy
--Multi-hit that can launch opponents.
--last hitbox will probably whiff if the move is blocked

just to let you know these are are the ONLY moves Gunther has data for. if anyone gets this character to be playable to test his moves so far keep that in mind cause im gonna assume that game won't like it when you do a move that doesn't exist.