Touched up camera movements
Made it so that the camera shake pauses when the game is paused.
UI HUD now stays in place while camera is shaking.
Russel's stage background is set to tile in all directions. This is done to reset the draw buffer so we don't see previous frames while the camera is shaking.