Edited Rus 6S
First Hit: 10 -> 8 damage
Enhancer now transitions immediately

Done to make Russel's other attacks more useful. This move should be good as a long-ranged conversion tool, but not so strong as to be the only thing he cancel's into.
As for the enhancer, I wanna see if requiring the player to time their enhance input would make the move more interesting.