Made adjustements to Bev's Backdash
I've hardcoded the game to set Bev's traction to a lower value during her backdash, and gave her backdash an extra step back. This temporarily done by editing the code for oPlayerController, in the future we should add an extra variable in the character editor so that we can define a character's backdash traction, and if/when they get a second step to their momentum.