Tech rolls now have a small vulnerability window
Tech rolls are vulnerable for 4 frames near the end of their animation. This is to give a small amount of added risk to the option, as tech rolls allow you to roll behind the attacker and potentially mount a counter-attack, in addition to avoiding meaty setups. It also creates a distinction between tech roll and quick getup, as the latter does not have this vulnerability, meaning you can either risk getting in or out, or stay in place to be safe.

Also fixed a bug where jumping before a quick getup would "store" the jump angle, and holding up afterwards would use the previously stored angle.