Experimented with the RC activation
The EX flash no longer appears on the Rush Cancel Startup. This is to improve clarity, since there's a lot of yellow on the screen due to RC bars at the top and bottom.

I moved the RC banners to be at a lower depth so that the top one would display below the player.