Added Air/Ground only enhancers
To implement this feature, I created a new variable called "requiredPosition." It is very similar to the "enhanced" variable, as it too stores an array to reference each enhancer.

I need to do it this way because the PerformMotionInputs script does not know which attack's special data to pull from, instead it relies on the SetMotionInputs script to set a bunch of variables that describe the various properties the special move we're performing. requiredPosition must be an array because otherwise we wouldn't know which enhancer to check the required position for.

Also fixed a bug where Beverly's 623 enhancer was usable even when below 25 meter.