Fixed input end crash
When the game checks to see if it should be handling motion inputs, it first checks to see if you are in any "special move" state. If not, then all the value pertaining to motion inputs get reset

If you are in a special state, there was an extra check to see if you were at the end of the input window. Since the input window is now an array, there's no way to know which one is the last, since they could theoretically be in any order.

If you are not at the end of the input window, then the game processes motion inputs.

To fix the issue, I simply got rid of the extra check for the input window end. There's no real need to reset everything before the move ends, since it'll get reset anyways, and as far as I have tested, removing this check has not resulted in any weird behavior.