Fixed an issue with unblockables
If you got hit by an attack during hitstop (not just a projectile, but any attack in general), then you would be unable to block.

This is because the game was setting the player's "canBlock" var to false during hitstop. Projectiles don't experience the same amount of hitstop as players, they always go through 10 frames of hitstop no matter what, so if the projectile had more than 10 frames of hitstop in its attack data, it would cause the player to block the first hit, but then they couldn't block the second hit because they were in histop.

This has been fixed by having the canBlock value persist between states. Most states set it to true or false, but the hitstop and screen freeze states don't effect the value. This means that if you could block before hitstop, you'll be able to block during it and vice versa.

As a result of this change, autogaurd (if you are in blockstun, you are forced to block attacks) has been temporarily removed.