Began reworking hitByGroups
I noticed that multiple players were unable to attack the same object at the same time. This is because each object only kept track of the group IDs that hit it, not who hit it.

This basically means I have to rework hitbox groups. I've managed to get it working by giving each player a list of objects that its hit, and a struct of object IDs that have hit it. Objects will be able to handle group values separately and will be able to get hit by multiple objects at once.