Adjusted Jay's Standing heavy
Increased Startup
14f -> 16f

Decreased Safety on Block
-3f -> -11f

Decreased the range of the hitbox slightly.

I don't want Jay to have a very safe poke that hits the opponent crouching. Jay is a tricky and evasive character, so he should be encouraged to attack in unorthodox ways. He has some other tools to compensate.