Fixed a bug with Bev's Super
If Bev uses this super, she no longer causes the DANGER text to get stuck if she is interrupted during the startup.

I just made the startup of the move invincible during the DANGER text, but the move is not a reversal. She can still be interrupted before the hitbox comes out. The projectile will always spawn, even if she gets hit.