Added No-Numpad Mode
In the event that the player's keyboard does not have a numpad, this mode allows them to play as P2 (albeit roughly, as there's not a lot of space on the keyboard.)

This is kind-of a temporary solution.

P2's menu keys get set to the 0-9 keys on the keyboard. They still use the arrow keys to move.