Began SC refactor
I'm gunna be restructuring SC's functionality to make him easier to maintain as new features get added to the game. I'll start by making it so that there's only ever one instance of a spirit that's active at all times. Right now, the game creates and destroys instances of SC as it needs to, which causes some issues.

I've made it so that the game will quickly boot in a match with 2 Jays so I don't have to wait through the character select.