Began adding spirit States
I've decided to give spirits their own state machine, since they only need a to do a few things.

For players, I've renamed spiritState to SpiritON, since that variable is meant to track whether they are in spiritON or not. This should be replaced in every instance of spiritState, but I may have missed some, we'll see.