Spirit OFF attacks that activate ON work now
If an attack that can be done with a spirit in Spirit OFF puts you in Spirit ON, it will now work properly.

Also fixed an issue where recovering from a spirit break in the middle of an attack would cause your spirit to come back and also perform the attack.