Remote Mode now functional
After stewing on it for a while, I have decided to allow SC to move and attack through walls. This is the easiest solution I could think of because there's just so many different scenarios that can happen where Jay and SC end up on different platforms with different heights and one's on the ground and one isn't and yada yada yada. SC is also a ghost, so IMO it makes sense that he can phase through walls.

Also the spawn offset for moves that use SC in Spirit OFF now works properly.