Spirit now moves on its own when host is cornered
If the host tries to move into a wall, the spirit will now move on its own while the host stays in place. This is to make controlling the spirit feel more intuitive, as even though the host is cornered, they probably still expect the spirit to move on its own. This doesn't work with dashes, however, so I'll need to think of a workaround.

Also began adding some states relating to grabs and throws for Jay and SC.