Began work on Spirit RC's
Spirits will now disappear when their host does a Rush Cancel while performing a move in Spirit OFF.

Actually, RC's work pretty much perfectly with spirits. The only thing that feels off is when the host and the spirit are on opposite sides of the opponent, where both the host and the spirit to move in the same direction direction despite facing opposite ways, which isn't ideal since you'd want the spirit to move towards the opponent instead of away from them.