Updated Russel's Enh. Down Special
Decreased total duration
57f -> 43f

Increased safety on block
-20f -> +4f

This move now travels forwards a bit during the startup.

This is to allow it to connect more reliable after the regular down special, and it keeps Russel closer to the opponent. I also decided to experiment with making the enhanced version plus on block to allow Russel a way to spend meter to keep pressure going.

(Also reverted some debug code)