Began work on camera behavior + Russel's Shorthop
I'm changing how the camera works in singleplayer to make it more useful for platforming, since putting Russel in the center of the screen makes it harder to see any obstacles in front of him.

I also increased the height of Russel's shorthop. It can now clear small ledges that it couldn't before. This shouldn't really change how effective it is in VS mode, since he still cannot cross-up opponents with it.

Lastly, I started experimenting with different level layouts.