More progress on camera functionality
It kinda works, but still need to iron out some of the kinks. It interacts weirdly with inputting a backwards side special, which is something I'll have to fix.

I also discovered a bug with the landing buffer. If you perform an attack in mid-air, then land and buffer an attack, the hit and hurtboxes from the aerial attack will persist until the grounded attack ends.

Also you can't landing buffer special moves apparently? Wonder what that's all about.