Side specials can now be turned around
In singleplayer, you can now turn around with your side special in the air or during a combo. This makes it a lot more intuitive and flexible, as you can use it to correct your trajectory in the air, or suddenly hit something behind you.

Also adjusted some of the camera behavior a bit more, still needs more tuning.