More progress on Russel Stage 1
Almost got the blockout done, just a couple more screens to go.

Right now, this level has about 2 minutes of content. When it's complete, it'll likely have 3 minutes, which is pretty good for a first level.

There's also a secret in this level. If you run away from the second rolling object (white circle), eventually it'll open up a hole in the floor that leads to a secret room.