Tested the game's performance
I left the game running in singleplayer for about 15-17 minutes. FPS was pretty stable at around 1800 fps.

I tested the campaign tutorial level and got 1400 fps while the ConstructionGuyNPC was loaded, but then shot back up to 2000 once he was KO'd.

I put 3 of them into the level and the fps tanked to 800. This means that there's some optimizations that need to happen in the oPlayerController script.

The new level I blocked out yesterday has 22 enemies total. For sure going to have to optimize this stuff. I have some ideas for how to do it, like only running the step event when the enemies are close to the player, or de-loading them once the player is too far away.