Began planning out the overworld map
I tried making a quick layout for the overworld, but I've got some creative block and I don't really have a lot of ideas right now.

I also experimented with putting the camera a little above the player to allow them to see what's ahead a little better, since the player is going to be traveling upwards as they progress through the arcade.