Got the rough draft for the overworld in
It's still missing Minigame #3, since I had a different idea for what to make it (although I completely forgot what it was : / )

Fight 2 and Minigame 2 are both optional levels which take place in Beverly's Jazzercise Studio. These are unlocked after the 2nd platforming level and are accessed via a set of stairs.