Experimented with putting enemy AIs in the level
Right now, I'm using Russels with the black alt to represent the basic enemies.

I've gotten one hooked up so far and it works fairly well. A bit rudimentary, but it works. It does require me to place and link up 3 different objects per enemy with no prefabs, but perhaps down the line I'll be able to create an object that sets it up automatically in the future.

Now I need to create the unique AI for the first enemy. This is a very basic enemy that only occasionally moves and tries to hit the player with a 3-hit jab combo. I've named this enemy "Gin."