Got the enemy spawner object working
Instead of manually placing 3 objects in the level for a single enemy, I've created an object which will create those objects for me, which will make placing enemies in levels much easier/faster.

This object can also track the health of it's spawned enemy and destroy it when necessary. It also activates/deactivates once the player gets within range.