Began experimenting with minigame level
I started designing the first minigame level, which happens after the first platforming level.

There's a lot of work that needs to be done before this level can be considered playable. For one, we don't have a way of moving the camera vertically at the moment, so that will need to be addressed.

All of the pop-up assets such as the combo counter, Round 1 text, counter hit text, etc. will have to be adjusted to compensate.

We also need to implement moving semi-solid platforms for this next level to work.