Enemies can no longer hit each other + CG fix
Players now have a "teamID" which tells the game which team they belong to. If two players are not on the same team, they are allowed to hit each other.

Also fixed a bug with command/hit grabs where the game would crash if the player grabbed an opponent out of an attack with a hitbox active.