Began working on Room Transition object
This is the object that can transport the player in between different rooms of a level.

Some more work will have to be done in order to get the timer and money to carry over, since the timer and money data is in a format that is hard to use.