Discovered something unfortunate with place_meeting
I was trying to make the collisions in this game more flush with walls, however I figured out that the place_meeting function does not work if two objects are overlapping by less than half a pixel. This means true flush collision is going to require a custom collision function which uses the bounding boxes of objects.

I don't have the time to make that right now, so I'll save it for later.