Began work on pushback rework
The current knockback/pushback system is hard to balance around, since each player moves back a specific amount. Moreover, the specified knockback amount could only be in integers, and the amount of knockback recieved would grow exponentially as the number grew higher. (e.g., a knockback of 4 would move the opponent 4+3+2+1 pixels back, and 5 would be 5+4+3+2+1, etc.)

I'm now making it so that the knockback parameter now specifies the distance the opponent travels instead of the starting velocity.