Added run acceleration
Characters now accelerate when they start running.

This is to make running feel more natural, and it helps prevent looping dash-in pressure from mindlessly holding the run button. It does make a few combos a lot harder though, I'll play around with it.

Also, remember to add this to the Character editor at some point.