Fixed issue with controllerNotifyList
Added an extra check to the char sel to see if its ID is already contained in the controllerUpdateNotifyList.

Since the char sel is an innate object of the room its in, it will always have the same ID every time. This check ensures that it does not add itself to that list multiple times.