Jay's teleport now functional
You can input left or right to teleport to that side of the screen.

Jay will always jump towards the opponent.

No enhancer yet, but doing half circle back during the move will cause Jay to plummet down to the floor directly above the opponent with Smooth Criminal and activate Spirit ON.

The mixups this move has is insane lmao.