More work on Arcade Stage
I hate it, but I can't really spend any more time on it. It's a lot easier to see the characters on this stage, but it also doesn't really match the art style. I might re-do it again in the future, but I think I'd rather do other stages first.

Also gave this stage a ceiling to prevent characters from going too high.