Adjusted Jay 5.H
Improved frame advantage on block
-11f -> -9f

This move no longer hops over lows.
Added an extra hurtbox during the moves recovery.

I thought giving Jay a low crush like this was too much. It was also hard to punish this move sometimes, so I added an extra hurtbox on the recovery. This move is still safe if used from a distance.