Investigating bug where frame advantage counter is wrong
It appears that the player is able to block on the exact same frame that their hitstun reaches 0. This means that even though a move is +7 on hit, in reality it's only +6 since the opponent is able to block on the 7th frame. They can't do anything else tho. I'm trying to see if I can find an elegant way to fix this.