Player can no longer block on last frame of hitstun
Fixed an issue where player's would be able to block/act on the last frame of hitstun, meaning that combos that should be 1-frame links on paper simply didn't work.

There's now a new issue where the game won't recognize 1-frame link combos if the first move is the first time the player hits their opponent.